// src/app/models/cong-viec/xu-ly-xong-viec.ts
export interface XuLyXongViec {}

export interface TimKiemXuLyCongViec {
    noiDung?: string;
    congViecPhatSinh?: number;
    nam?: number;
    thang?: number;
    userId?: number;
    phongBanId?: number;
    donViId?: number;
    vaiTro?: number;
    trangThai?: number;
    timChinhXac?: number;
}